hook syscall in libfstack.a
