+++ app/nginx-1.11.10/src/event/modules/ngx_select_module.c
@@ -72,12 +72,6 @@ ngx_module_t  ngx_select_module = {
     NGX_MODULE_V1_PADDING
 };
 
-#if (NGX_HAVE_FSTACK)
-#define FF_FD_BITS 16
-#define CHK_FD_BIT(fd)          (fd & (1 << FF_FD_BITS))
-#define CLR_FD_BIT(fd)          (fd & ~(1 << FF_FD_BITS))
-#endif
-
 static ngx_int_t
 ngx_select_init(ngx_cycle_t *cycle, ngx_msec_t timer)
 {
@@ -155,16 +149,16 @@ ngx_select_add_event(ngx_event_t *ev, ngx_int_t event, ngx_uint_t flags)
 
     if (event == NGX_READ_EVENT) {
 #if (NGX_HAVE_FSTACK)
-        if (CHK_FD_BIT(c->fd))
-            FD_SET(CLR_FD_BIT(c->fd), &master_read_fd_set);
+        if (FF_FD_CHK(c->fd))
+            FD_SET(FF_FD_CLR(c->fd), &master_read_fd_set);
 #else
         FD_SET(c->fd, &master_read_fd_set);
 #endif
 
     } else if (event == NGX_WRITE_EVENT) {
 #if (NGX_HAVE_FSTACK)
-        if (CHK_FD_BIT(c->fd))
-            FD_SET(CLR_FD_BIT(c->fd), &master_write_fd_set);
+        if (FF_FD_CHK(c->fd))
+            FD_SET(FF_FD_CLR(c->fd), &master_write_fd_set);
 #else
         FD_SET(c->fd, &master_write_fd_set);
 #endif
@@ -203,16 +197,16 @@ ngx_select_del_event(ngx_event_t *ev, ngx_int_t event, ngx_uint_t flags)
 
     if (event == NGX_READ_EVENT) {
 #if (NGX_HAVE_FSTACK)
-       if (CHK_FD_BIT(c->fd))
-            FD_CLR(CLR_FD_BIT(c->fd), &master_read_fd_set);
+       if (FF_FD_CHK(c->fd))
+            FD_CLR(FF_FD_CLR(c->fd), &master_read_fd_set);
 #else
         FD_CLR(c->fd, &master_read_fd_set);
 #endif
 
     } else if (event == NGX_WRITE_EVENT) {
 #if (NGX_HAVE_FSTACK)
-        if (CHK_FD_BIT(c->fd))
-            FD_CLR(CLR_FD_BIT(c->fd), &master_write_fd_set);
+        if (FF_FD_CHK(c->fd))
+            FD_CLR(FF_FD_CLR(c->fd), &master_write_fd_set);
 #else
         FD_CLR(c->fd, &master_write_fd_set);
 #endif
@@ -345,7 +339,7 @@ ngx_select_process_events(ngx_cycle_t *cycle, ngx_msec_t timer,
 
         if (ev->write) {
 #if (NGX_HAVE_FSTACK)
-            if (FD_ISSET(CLR_FD_BIT(c->fd), &work_write_fd_set)) {
+            if (FD_ISSET(FF_FD_CLR(c->fd), &work_write_fd_set)) {
 #else
             if (FD_ISSET(c->fd, &work_write_fd_set)) {
 #endif
@@ -356,7 +350,7 @@ ngx_select_process_events(ngx_cycle_t *cycle, ngx_msec_t timer,
 
         } else {
 #if (NGX_HAVE_FSTACK)
-            if (FD_ISSET(CLR_FD_BIT(c->fd), &work_read_fd_set)) {
+            if (FD_ISSET(FF_FD_CLR(c->fd), &work_read_fd_set)) {
 #else
             if (FD_ISSET(c->fd, &work_read_fd_set)) {
 #endif
