+++ app/redis-3.2.8/src/Makefile.dep
@@ -5,7 +5,6 @@ ae_evport.o: ae_evport.c
 ae_kqueue.o: ae_kqueue.c
 ae_ff_kqueue.o: ae_ff_kqueue.c
 ae_select.o: ae_select.c
-anet_ff.o: anet_ff.c anet_ff.h
 anet.o: anet.c fmacros.h anet.h
 aof.o: aof.c server.h fmacros.h config.h solarisfixes.h \
  ../deps/lua/src/lua.h ../deps/lua/src/luaconf.h ae.h sds.h dict.h \
@@ -135,7 +134,7 @@ server.o: server.c server.h fmacros.h config.h solarisfixes.h \
  ../deps/lua/src/lua.h ../deps/lua/src/luaconf.h ae.h sds.h dict.h \
  adlist.h zmalloc.h anet.h ziplist.h intset.h version.h util.h latency.h \
  sparkline.h quicklist.h zipmap.h sha1.h endianconv.h crc64.h rdb.h rio.h \
- cluster.h slowlog.h bio.h asciilogo.h anet_ff.h
+ cluster.h slowlog.h bio.h asciilogo.h
 setproctitle.o: setproctitle.c
 sha1.o: sha1.c solarisfixes.h sha1.h config.h
 slowlog.o: slowlog.c server.h fmacros.h config.h solarisfixes.h \
