--- app/redis-3.2.8/src/anet_ff.h
@@ -1,7 +1,0 @@
-#ifndef __ANET_FF_H__
-#define __ANET_FF_H__
-
-void ff_mod_init();
-
-#endif /* ifndef __ANET_FF_H__ */
-
