+++ example/main_epoll.c
@@ -6,13 +6,13 @@
 #include <strings.h>
 #include <sys/types.h>
 #include <sys/socket.h>
+#include <sys/epoll.h>
 #include <arpa/inet.h>
 #include <errno.h>
 #include <assert.h>
 
 #include "ff_config.h"
 #include "ff_api.h"
-#include "ff_epoll.h"
 
 
 #define MAX_EVENTS 512
@@ -20,7 +20,6 @@
 struct epoll_event ev;
 
 struct epoll_event events[MAX_EVENTS];
-struct kevent kqevents[MAX_EVENTS];
 
 int epfd;
 int sockfd;
@@ -61,38 +60,43 @@ int loop(void *arg)
 {
     /* Wait for events to happen */
 
-    int nevents = ff_epoll_wait(epfd,  events, MAX_EVENTS, 0);
+    int nevents = epoll_wait(epfd,  events, MAX_EVENTS, 0);
     int i;
 
-    for (i = 0; i < nevents; ++i) {	
+    for (i = 0; i < nevents; ++i) {    
         /* Handle new connect */
         if (events[i].data.fd == sockfd) {
-            int nclientfd = ff_accept(sockfd, NULL, NULL);
+            int nclientfd = accept(sockfd, NULL, NULL);
             assert(nclientfd > 0);
             /* Add to event list */
-	    ev.data.fd = nclientfd;
-	    ev.events  = EPOLLIN;
-	    assert(ff_epoll_ctl(epfd, EPOLL_CTL_ADD, nclientfd, &ev) == 0);
+            ev.data.fd = nclientfd;
+            ev.events  = EPOLLIN;
+            assert(epoll_ctl(epfd, EPOLL_CTL_ADD, nclientfd, &ev) == 0);
             //fprintf(stderr, "A new client connected to the server..., fd:%d\n", nclientfd);
+
         } else { 
             if (events[i].events & EPOLLERR ) {
                 /* Simply close socket */
-		ff_epoll_ctl(epfd, EPOLL_CTL_DEL,  events[i].data.fd, NULL);
-                ff_close(events[i].data.fd);
+                epoll_ctl(epfd, EPOLL_CTL_DEL,  events[i].data.fd, NULL);
+                close(events[i].data.fd);
                 //fprintf(stderr, "A client has left the server...,fd:%d\n", events[i].data.fd);
+
             } else if (events[i].events & EPOLLIN) {
                 char buf[256];
-                size_t readlen = ff_read( events[i].data.fd, buf, sizeof(buf));
-                //fprintf(stderr, "bytes are available to read..., readlen:%d, fd:%d\n", readlen,  events[i].data.fd);
-    		if(readlen > 0){
-                    ff_write( events[i].data.fd, html, sizeof(html));
-		} else {
-		    ff_epoll_ctl(epfd, EPOLL_CTL_DEL,  events[i].data.fd, NULL);
-    		    ff_close( events[i].data.fd);
-                    //fprintf(stderr, "A client has left the server...,fd:%d\n", events[i].data.fd);		
-		}
+                size_t readlen = read( events[i].data.fd, buf, sizeof(buf));
+                //fprintf(stderr, "bytes are available to read..., readlen:%d, fd:%d\n", readlen, events[i].data.fd);
+
+                if(readlen > 0){
+                    write( events[i].data.fd, html, sizeof(html));
+
+                } else {
+                    epoll_ctl(epfd, EPOLL_CTL_DEL,  events[i].data.fd, NULL);
+                    close( events[i].data.fd);
+                    //fprintf(stderr, "A client has left the server...,fd:%d\n", events[i].data.fd);        
+                }
+
             } else {
-                fprintf(stderr, "unknown event: %8.8X\n", events[i].events);
+                //fprintf(stderr, "unknown event: %8.8X\n", events[i].events);
             }
         }
     }
@@ -109,13 +113,13 @@ int main(int argc, char * argv[])
 
     ff_init(conf, argc, argv);
 
-    sockfd = ff_socket(AF_INET, SOCK_STREAM, 0);
-    printf("sockfd:%d\n", sockfd);
-    if (sockfd < 0)
-        printf("ff_socket failed\n");
+    sockfd = socket(AF_INET, SOCK_STREAM, 0);
+    if (sockfd < 0) {
+        printf("socket failed\n");
+    }
 
     int on = 1;
-    ff_ioctl(sockfd, FIONBIO, &on);
+    ioctl(sockfd, FIONBIO, &on);
 
     struct sockaddr_in my_addr;
     bzero(&my_addr, sizeof(my_addr));
@@ -123,20 +127,20 @@ int main(int argc, char * argv[])
     my_addr.sin_port = htons(80);
     my_addr.sin_addr.s_addr = htonl(INADDR_ANY);
 
-    int ret = ff_bind(sockfd, (struct linux_sockaddr *)&my_addr, sizeof(my_addr));
+    int ret = bind(sockfd, (struct sockaddr *)&my_addr, sizeof(my_addr));
     if (ret < 0) {
-        printf("ff_bind failed\n");
+        printf("bind failed\n");
     }
 
-    ret = ff_listen(sockfd, MAX_EVENTS);
+    ret = listen(sockfd, MAX_EVENTS);
     if (ret < 0) {
-        printf("ff_listen failed\n");
+        printf("listen failed\n");
     }
-	
-    assert((epfd = ff_epoll_create(0)) > 0);
+    
+    assert((epfd = fepoll_create(0)) > 0);
     ev.data.fd = sockfd;
     ev.events = EPOLLIN;
-    ff_epoll_ctl(epfd, EPOLL_CTL_ADD, sockfd, &ev);
+    epoll_ctl(epfd, EPOLL_CTL_ADD, sockfd, &ev);
     ff_run(loop, NULL);
     return 0;
 }
