+++ lib/Makefile
@@ -128,8 +128,7 @@ FF_SRCS+=			\
 	ff_syscall_wrapper.c \
 	ff_subr_prf.c    \
 	ff_vfs_ops.c  \
-	ff_veth.c  \
-	ff_epoll.c
+	ff_veth.c
 
 FF_HOST_SRCS+= \
 	ff_host_interface.c \
@@ -138,6 +137,7 @@ FF_HOST_SRCS+= \
 	ff_dpdk_if.c \
 	ff_dpdk_kni.c \
 	ff_dpdk_pcap.c \
+	ff_epoll.c \
 	ff_init.c
 
 ifdef FF_IPSEC
