+++ lib/ff_epoll.c
@@ -1,34 +1,24 @@
-#include <sys/param.h>
-#include <sys/limits.h>
-#include <sys/uio.h>
-#include <sys/proc.h>
-#include <sys/syscallsubr.h>
-#include <sys/module.h>
-#include <sys/param.h>
-#include <sys/malloc.h>
-#include <sys/socketvar.h>
-#include <sys/event.h>
-#include <sys/kernel.h>
-#include <sys/refcount.h>
-#include <sys/sysctl.h>
-#include <sys/pcpu.h>
-#include <sys/select.h>
-#include <sys/poll.h>
-#include <sys/event.h>
-#include <sys/file.h>
+#include <stdio.h>
+#include <stdint.h>
+#include <string.h>
+#include <stdlib.h>
+#include <stdarg.h>
+#include <sched.h>
+#include <fcntl.h>
+#include <errno.h>
+#include <assert.h>
+#include <unistd.h>
 #include <netinet/in.h>
-#include <netinet/tcp.h>
-#include <sys/ttycom.h>
-#include <sys/filio.h>
-#include <sys/sysproto.h>
-#include <sys/fcntl.h>
-#include <machine/stdarg.h>
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <sys/time.h>
+#include <sys/select.h>
+#include <sys/syscall.h>
+#include <arpa/inet.h>
+#include <sys/epoll.h>
 
 #include "ff_api.h"
-#include "ff_epoll.h"
 #include "ff_errno.h"
-#include "ff_host_interface.h"
-
 
 
 int
@@ -37,16 +27,16 @@ ff_epoll_create(int size __attribute__((__unused__)))
 	return ff_kqueue();
 }
 
-
 int 
 ff_epoll_ctl(int epfd, int op, int fd, struct epoll_event *event)
 {
+	struct kevent kev[3];
+
 	if (!event && op != EPOLL_CTL_DEL) {
-        ff_os_errno(ff_EINVAL);
+        errno = EINVAL;
 		return -1;
 	}
 
-	struct kevent kev[3];
 	if (op == EPOLL_CTL_ADD){
 		EV_SET(&kev[0], fd, EVFILT_READ,
 			EV_ADD | (event->events & EPOLLIN ? 0 : EV_DISABLE), 0, 0, NULL);
@@ -66,7 +56,7 @@ ff_epoll_ctl(int epfd, int op, int fd, struct epoll_event *event)
 		EV_SET(&kev[2], fd, EVFILT_USER, 0,
 		    NOTE_FFCOPY | (event->events & EPOLLRDHUP ? 1 : 0), 0, NULL);		
 	} else {
-		ff_os_errno(ff_EINVAL);
+		errno = EINVAL;
 		return -1;
 	}
 
@@ -76,26 +66,30 @@ ff_epoll_ctl(int epfd, int op, int fd, struct epoll_event *event)
 int 
 ff_epoll_wait(int epfd, struct epoll_event *events, int maxevents, int timeout)
 {
+    int i, ret;
+    struct kevent *evlist; 
+	unsigned int event_one = 0;
+
 	if (!events || maxevents < 1) {
-		ff_os_errno(ff_EINVAL);
+		errno = EINVAL;
 		return -1;
 	}
 	
-	struct kevent *evlist = malloc(sizeof(struct kevent)*maxevents, M_DEVBUF, M_ZERO|M_NOWAIT);
+	evlist = malloc(sizeof(struct kevent) * maxevents);
 	if(NULL == evlist){
-		ff_os_errno(ff_EINVAL);
+		errno = EINVAL;
 		return -1;		
 	}
-	memset(evlist, 0, sizeof(struct kevent)*maxevents);
-	
-	int ret = ff_kevent(epfd, NULL, 0, evlist, maxevents, NULL);
+
+	ret = ff_kevent(epfd, NULL, 0, evlist, maxevents, NULL);
 	if (ret == -1) {
-		free(evlist, M_DEVBUF);
+		free(evlist);
 		return ret;
 	}
 
-	unsigned int event_one = 0;
-	for (int i = 0; i < ret; ++i) {
+	memset(evlist, 0, sizeof(struct kevent) * ret);
+
+	for (i = 0; i < ret; ++i) {
 		event_one = 0;
 		if (evlist[i].filter & EVFILT_READ) {
 			event_one |= EPOLLIN;
@@ -115,7 +109,7 @@ ff_epoll_wait(int epfd, struct epoll_event *events, int maxevents, int timeout)
 		events[i].data.fd  = evlist[i].ident;
 	}
 	
-	free(evlist, M_DEVBUF);
+	free(evlist);
 	return ret;
 }
